Fix make_ktbs behavior to make a unique root_uri
This is needed if multiple kTBS instances run at the same time and not using different URI (case were the kTBS are not run on http for example). With the old behavior, the kTBS instances could share locks for different resources. This commit fix this issue.

If multiple kTBS instances run on IOMemory then the PID of the kTBS are added to the root_uri to differentiate them.