rdfrest: added general support for JSON-LD
NB: for the moment,
we keep the previous JSON-LD serializers/parsers for ktbs,
even if that means redundant dependencies.

ktbs serializers are specialized for each class, so we need them anyway.
The ktbs parser uses a custom document loader
to avoid loading the kTBS context from the network,
and rdflib-jsonld does not seem to be able to do that...