fix #78 and improve JSON API
* no longer generates both `k:hasBase` and `k:contains` for subbases
* no longer requires `inBase` when posting subbase in simple JSON

This is achieved through a slight hack:
the JSON parser (which does not know the target of the graph)
always adds `inRoot` to a base,
if neither `inRoot` nor `inBase` are present.
Then, the base receiving the subvase via POST replaces any
`k:hasBase` by `k:contains`.

An alternative solution (may be in a future version)
would be to change the kTBS vocabulary,
and use the same predicate `k:contains`
even between the root and the 1st level bases.
But this would create backward incompatibilities.